requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
PyMySQL==1.1.1
cryptography==44.0.1
